Add a My Location parameter to the Google Maps plugin (#898)
Introduces a My Location parameter to the Google Maps plugin. This is a feature that Google Maps offers that puts a blue location dot on the user's location, and provides a clickable compass. This parameter works the same way as the zoom, or scale options, and can be enabled or disabled by updating the map option.