[google_sign_in] Make the Delegate non-final to allow overriding (#2660)
Make the Delegate non-final to allow overriding. IDelegate is not enough for v2 embedding for GoogleSignIn and 1p auth plugin because activity and activity binding have to be involved. If we keep Delegate as a field for those plugins, make Delegate non-final so that it's easier for tests to override.