[video_player] Add URL for exoplayer to maven for all dependents (#3979)
JCenter has the older version of ExoPlayer that the plugin currently uses, but Google's maven server doesn't so we need to add a workaround in the short term for the removal of JCenter in #3924 to avoid breaking clients of video_player.

This was not caught in presubmit because the video_player example had this same workaround left over from an issue several years ago that caused JCenter not to have that version either; that issue was since fixed, but the workaround was never removed. This removes the example-level workaround so the example won't mask exoplayer maven issues in the future.

(Also temporarily re-adds the workaround to the camera example, which broke post-publishing, to fix the tree. All example workarounds should be removable once this change is published.)