[plugin_platform_interface] Fix 'verify' for Fakes (#4651)
The new `verify` method worked with `Mock` (since `_instanceToken`
returns a stub function in that case), but not `Fake` or a manual fake
since it was unconditionally accessing `_instanceToken`. This fixes that
new check to be gated behind the `MockPlatformInterfaceMixin` test, and
adds a unit test that covers `Fake` so this can't regress in the future.

Fixes post-publish breakage in the tree.