[webview_flutter] Remove duplicate integration test (#4666)
This test is for a change that was entirely in Android Java code when it
was written, so is testing behavior that is internal to the Android
implementation, and a copy of this test already runs
webview_flutter_android. Given that it's testing internal Android
behavior, only that copy is needed.

For unknown reasons, this copy seems to hang test runs, while the
webview_flutter_android copy does not. Since it's a duplicate that we
want to remove anyway, this simply removes it to fix the tree without
meaningful loss of coverage.

Also fixes a race condition in an integration test, where a future
could be completed more than once.

Part of https://github.com/flutter/flutter/issues/91004