[video_player] Federate mobile implementations (#4712)
Moves the iOS and Android implementations to their own federated packages. The iOS implementation package uses the name `avfoundation` rather than `ios` since it is likely that a macOS implementation could share a significant amount of code.

As part of the federation, migrates the mobile implementations to using in-package method channels rather than the shared method channel. Currently these are just duplicates of the existing method channels; updating them to new version of Pigeon will be done in a separate PR later.

Temporarily marks `video_player` as unpublishable to allow the implementations to be moved, rather than copied and deleted, in order to better preserve git history. A follow-up PR will restore it to publishable form.

No version change: The changes to `video_player_platform_interface` only affect development of the plugin, not use of the plugin.

Part of https://github.com/flutter/flutter/issues/68498
Part of https://github.com/flutter/flutter/issues/94224